LibWeb: Split BorderStyleValue out of StyleValue.{h,cpp}
